Use Atomic Design for components
