+++ package.json
@@ -52,7 +52,7 @@
     "gatsby-remark-default-html-attrs": "^1.0.0",
     "gatsby-remark-prismjs": "3.2.3",
     "gatsby-source-filesystem": "^2.1.1",
-    "gatsby-source-google-docs": "^1.14.0",
+    "gatsby-source-google-docs": "^1.14.2",
     "gatsby-transformer-remark": "^2.2.2",
     "metal-clipboard": "2.0.1",
     "metal-tabs": "2.3.1",
@@ -71,6 +71,7 @@
   "devDependencies": {
     "gatsby-plugin-force-trailing-slashes": "^1.0.4",
     "gatsby-plugin-remove-trailing-slashes": "^2.1.4",
+    "gatsby-plugin-resolve-src": "^2.0.0",
     "gh-pages": "2.0.1",
     "globby": "9.2.0",
     "rimraf": "2.6.3"
