--- src/components/Auth/index.js
@@ -1,3 +1,0 @@
-import Auth from './Auth';
-
-export default Auth;
