--- src/components/CodeClipboard/index.js
@@ -1,3 +1,0 @@
-import CodeClipboard from './CodeClipboard';
-
-export default CodeClipboard;
