--- src/components/CodeTabs/index.js
@@ -1,3 +1,0 @@
-import CodeTabs from './CodeTabs';
-
-export default CodeTabs;
