--- src/components/DeployButton/index.js
@@ -1,3 +1,0 @@
-import DeployButton from './DeployButton';
-
-export default DeployButton;
