--- src/components/LayoutNav/index.js
@@ -1,3 +1,0 @@
-import LayoutNav from './LayoutNav';
-
-export default LayoutNav;
