--- src/components/LogoutContainer/index.js
@@ -1,3 +1,0 @@
-import LogoutContainer from './LogoutContainer';
-
-export default LogoutContainer;
