--- src/components/Sidebar/index.js
@@ -1,3 +1,0 @@
-import Sidebar from './Sidebar';
-
-export default Sidebar;
