--- src/components/SimpleFooter/index.js
@@ -1,3 +1,0 @@
-import SimpleFooter from './SimpleFooter';
-
-export default SimpleFooter;
