--- src/components/Timeline/index.js
@@ -1,3 +1,0 @@
-import Timeline from './Timeline';
-
-export default Timeline;
