--- src/components/Typography/index.js
@@ -1,3 +1,0 @@
-import * as Typography from './Typography';
-
-export default Typography;
