+++ src/components/atoms/index.js
@@ -1,0 +1,5 @@
+import Typography from "./Typography"
+
+export {
+	Typography
+}
