--- src/components/index.js
@@ -1,11 +1,0 @@
-import BlogList from './BlogList';
-import CodeClipboard from './CodeClipboard';
-import CodeTabs from './CodeTabs';
-import Footer from './Footer';
-import Icon from './Icon';
-import LayoutNav from './LayoutNav';
-import Sidebar from './Sidebar';
-import SimpleFooter from './SimpleFooter';
-import Typography from './Typography';
-
-export default {BlogList, CodeClipboard, CodeTabs, Footer, Icon, LayoutNav, Sidebar, SimpleFooter, Typography};
