+++ src/components/molecules/CodeClipboard/index.js
