+++ src/components/molecules/index.js
@@ -1,0 +1,17 @@
+import Auth from "./Auth"
+import CodeClipboard from "./CodeClipboard"
+import CodeTabs from "./CodeTabs"
+import DeployButton from "./DeployButton"
+import Login from "./Login"
+import LogoutContainer from "./LogoutContainer"
+import Search from "./Search"
+
+export {
+	Auth,
+	CodeClipboard,
+	CodeTabs,
+	DeployButton,
+	Login,
+	LogoutContainer,
+	Search
+}
