+++ src/components/organisms/SimpleFooter/index.js
