+++ src/components/organisms/index.js
@@ -1,0 +1,11 @@
+import Footer from "./Footer"
+import LayoutNav from "./LayoutNav"
+import Sidebar from "./Sidebar"
+import SimpleFooter from "./SimpleFooter"
+
+export {
+	Footer, 
+	LayoutNav,
+	Sidebar,
+	SimpleFooter
+}
