+++ src/components/templates/index.js
@@ -1,0 +1,5 @@
+import Articles from "./Articles"
+
+export {
+	Articles
+}
